Moved denoising strength to stateful widget for better slider
